Fixed declaration of t_output_last : real --> integer
Former-commit-id: b0ff7162ed9358135b605ef2292c7e2860d075c9

+++ IE/Ridley_serial/src/IE_ModIo.f90
@@ -24,7 +24,7 @@ module IE_ModIo
 
   integer             :: dn_output(MaxFile)=-1
   real                :: dt_output(MaxFile)=-1.0
-  real                :: t_output_last(MaxFile)=-1
+  integer             :: t_output_last(MaxFile)=-1
 
 contains
